# check unexpected high bits for invalid characters