--- a/lib/bn.js
+++ b/lib/bn.js
@@ -186,6 +186,7 @@
   function parseHex (str, start, end) {
     var r = 0;
     var len = Math.min(str.length, end);
+    var z = 0;
     for (var i = start; i < len; i++) {
       var c = str.charCodeAt(i) - 48;
 
@@ -206,9 +207,11 @@
         b = c;
       }
 
-      assert(c >= 0 && b <= 0xf, 'Invalid character');
       r |= b;
+      z |= b;
     }
+
+    assert(!(z & 0xf0), 'Invalid character in ' + str);
     return r;
   }