> 10% performance drop, but 200% better debuggability.

Hopefully this resolves that problem @indutny