@dcousens 2017-11-30
@indutny updated,  passes tests now