@indutny 2017-12-02
Sorry for delay. I just benchmarked it, and don't see any significant performance improvement. Do you see it on your side?

---

Benchmarks can be run with:
```sh
SELF_ONLY=1 node benchmarks/index.js create
```