@dcousens 2017-12-04
I measured an average 5-10% improvment... but the GC totally throws it off.