@fanatid 2017-12-05
node version: v9.2.0

master:
```
Seed: 82910582182c79d7d4d9592e15d53f0d
Benchmarking: create-10
bn.js#create-10 x 360,920 ops/sec ±2.15% (9 runs sampled)
------------------------
Fastest is bn.js#create-10
========================
Benchmarking: create-hex
bn.js#create-hex x 433,552 ops/sec ±5.14% (8 runs sampled)
------------------------
Fastest is bn.js#create-hex
========================
```

PR:
```
Seed: 5866d26a6ca6d6ce9eed6ba5f50ce0e6
Benchmarking: create-10
bn.js#create-10 x 359,505 ops/sec ±1.40% (9 runs sampled)
------------------------
Fastest is bn.js#create-10
========================
Benchmarking: create-hex
bn.js#create-hex x 459,686 ops/sec ±2.40% (9 runs sampled)
------------------------
Fastest is bn.js#create-hex
========================
```