@dcousens 2017-12-05
@fanatid how did you get consistent results :P