@dcousens 2018-02-01
@indutny good to merge?